+++ tweed5/serde-extension/src/main/java/de/siphalor/tweed5/serde/extension/api/read/result/TweedReadIssue.java
@@ -2,6 +2,7 @@ package de.siphalor.tweed5.serde.extension.api.read.result;
 
 
 import de.siphalor.tweed5.serde.extension.api.TweedReadContext;
+import de.siphalor.tweed5.serde.extension.api.path.EntryPath;
 import lombok.*;
 
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
@@ -10,13 +11,22 @@ public class TweedReadIssue {
 	@Getter
 	private final Exception exception;
 	private final TweedReadContext readContext;
+	@Getter
+	private final EntryPath entryPath;
+	@Getter
+	private final EntryPath valuePath;
 
 	public static TweedReadIssue error(String message, TweedReadContext readContext) {
-		return new TweedReadIssue(new Exception(message), readContext);
+		return error(new Exception(message), readContext);
 	}
 
 	public static TweedReadIssue error(Exception exception, TweedReadContext readContext) {
-		return new TweedReadIssue(exception, readContext);
+		return new TweedReadIssue(
+				exception,
+				readContext,
+				readContext.currentEntryPath(),
+				readContext.currentValuePath()
+		);
 	}
 
 	@Override